Refactor: Modal, Alert, RecommendCard Style 분리
