+++ src/components/Modal/RecommendCard/RecommendCard.jsx
@@ -1,94 +1,20 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
-import Modal from "./Modal";
-import IconMoreVertical from "../../assets/images/icon-more-vertical.svg";
-import ImgStar from "../../assets/images/star.svg";
+import Modal from "../Modal/Modal";
+import IconMoreVertical from "../../../assets/images/icon-more-vertical.svg";
 import axios from "axios";
 import { useLocation, useNavigate } from "react-router-dom";
-import RecommendEdit from "../Post/PostEdit/RecommendEdit";
-const RecommendDiv = styled.div`
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  height: 100%;
-  z-index: 999;
-  background-color: rgba(0, 0, 0, 0.3);
-  width: 100%;
-  max-width: 390px;
-`;
-
-const RecommendCardArticle = styled.article`
-  width: 304px;
-  background-color: white;
-  border-radius: 10px;
-  box-sizing: border-box;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-`;
-
-const RecommendListImg = styled.img`
-  width: 100%;
-  height: 180px;
-  object-fit: cover;
-  border-radius: 8px 8px 0 0;
-`;
-
-const RecommendTextSection = styled.section`
-  padding: 13px;
-  position: relative;
-`;
-
-const RecommendName = styled.h3`
-  font-size: 17px;
-  margin-bottom: 13px;
-  font-weight: 600;
-  display: inline-block;
-  width: 240px;
-  line-height: 20px;
-`;
-
-const RecommendScoreSpan = styled.span`
-  display: inline-block;
-  color: #000;
-  position: relative;
-  padding-left: 23px;
-  font-size: 15px;
-  font-weight: 600;
-  line-height: 16px;
-  &::after {
-    display: block;
-    content: "";
-    width: 16px;
-    height: 16px;
-    background-image: url(${ImgStar});
-    background-repeat: no-repeat;
-    background-size: cover;
-    position: absolute;
-    top: 45%;
-    left: 4px;
-    transform: translateY(-50%);
-  }
-`;
-
-const RecommendLocationP = styled.p`
-  font-size: 15px;
-  margin-bottom: 27px;
-  line-height: 17px;
-`;
-
-const RecommendMoreBtn = styled.button`
-  position: absolute;
-  top: 7%;
-  right: 1%;
-`;
-
-const RecommendCloseBtn = styled.button`
-  font-size: 14px;
-  font-weight: 600;
-`;
+import RecommendEdit from "../../Post/PostEdit/RecommendEdit";
+import {
+  RecommendDiv,
+  RecommendCardArticle,
+  RecommendListImg,
+  RecommendTextSection,
+  RecommendName,
+  RecommendScoreSpan,
+  RecommendLocationP,
+  RecommendMoreBtn,
+  RecommendCloseBtn,
+} from "./RecommendCardStyle";
 
 export default function RecommendCard({ cardClose, id, modalOpen }) {
   const location = useLocation();
@@ -159,13 +85,11 @@ export default function RecommendCard({ cardClose, id, modalOpen }) {
   return (
     <RecommendDiv>
       <RecommendCardArticle>
-        <h2 className="a11y-hidden">추천 맛집</h2>
+        <h3 className="a11y-hidden">추천 맛집 카드</h3>
         <RecommendListImg src={recommendInfo.itemImage} alt="" />
         <RecommendTextSection>
-          <RecommendName>
-            {recommendInfo.itemName}
-            <RecommendScoreSpan>{recommendInfo.price}.0</RecommendScoreSpan>
-          </RecommendName>
+          <RecommendName>{recommendInfo.itemName}</RecommendName>
+          <RecommendScoreSpan>{recommendInfo.price}.0</RecommendScoreSpan>
           <RecommendLocationP>{recommendInfo.link}</RecommendLocationP>
           <RecommendMoreBtn type="button" onClick={modalOpen}>
             <img src={IconMoreVertical} alt="더보기 아이콘" />
