+++ src/pages/Post/DetailPost.jsx
@@ -4,8 +4,8 @@ import PostItem from "../../components/Post/PostItem/PostItem";
 import Comment from "../../components/Comment/Comment";
 import styled from "styled-components";
 import Header from "../../components/common/Header/Header";
-import Modal from "../../components/Modal/Modal";
-import Alert from "../../components/Modal/Alert";
+import Modal from "../../components/Modal/Modal/Modal";
+import Alert from "../../components/Modal/Alert/Alert";
 import axios from "axios";
 import PostEdit from "../../components/Post/PostEdit/PostEdit";
 import BasicProfile from "../../assets/images/basic-profile-lg.svg";
